+++ CHANGELOG.md
@@ -170,6 +170,7 @@ CHANGELOG
 - Fix wrong operator for list-contains-value operation in sieve expert documentation (PR#2256 by Filip Pokorný).
 - Added documentation on `default_fields` parameter (PR#2293 by Filip Pokorný).
 - Updated documentation on `skip_header` parameter (PR#2313 by Filip Pokorný).
+- Viriback Unsafe Sites feed replaced with Viriback C2 Tracker. (PR#2266 by Filip Pokorný)
 
 ### Packaging
 - Remove deleted `intelmq.bots.experts.sieve.validator` from executables in `setup.py` (PR#2256 by Filip Pokorný).